@@ -1,7 +1,6 @@
-# ====================================== General Configuration =====================================
 [project]
 name = "mtmlda"
-version = "0.1.0"
+version = "1.0.0"
 description = "Markov Tree Multilevel Delayed Acceptance MCMC Library"
 readme = "README.md"
 license = {text = "MIT License"}
@@ -9,79 +8,38 @@ authors = [
     {name = "Maximilian Kruse", email = "maximilian.kruse@kit.edu"},
     {name = "Linus Seelinger", email = "linus.seelinger@kit.edu"},
 ]
-
-requires-python = ">=3.9"
+requires-python = ">=3.12"
 dependencies = [
-    "anytree>=2.12",
-    "numpy>=1.26"
+    "anytree>=2.12.1",
+    "numpy>=2.1.3",
 ]
 
-#---------------------------------------------------------------------------------------------------
 [project.optional-dependencies]
 model = [
-    "umbridge>=1.2.4"
-]
-postprocessing = [
-    "arviz>=0.17",
-    "pydot>=2.0",
-    "seaborn>=0.13",
-    "xarray>=2024.3"
+    "umbridge>=1.2.4",
 ]
 notebook = [
-    "jupyter>=1.0",
-    "pandas>=2.2"
+    "jupyter>=1.1.1",
+    "pandas>=2.2.3",
 ]
-style = [
-    "ruff>=0.3",
-    "black>=24.3",
-    "mypy>=1.9",
+postprocessing = [
+    "arviz>=0.20.0",
+    "pydot>=3.0.3",
+    "seaborn>=0.13.2",
+    "xarray>=2024.11.0",
+]
+development = [
+    "ruff>=0.8.1",
 ]
 
-#---------------------------------------------------------------------------------------------------
 [build-system]
-requires = ["setuptools>=58", "wheel"]
-build-backend = "setuptools.build_meta"
+requires = ["hatchling"]
+build-backend = "hatchling.build"
 
-
-
-
-# =============================================== Ruff =============================================
 [tool.ruff]
 line-length = 100
-src = ["src"]
-select = ["E", "F", "UP", "D",]
-ignore = []
-fixable = ["ALL"]
-unfixable = []
-exclude = [
-    ".bzr",
-    ".direnv",
-    ".eggs",
-    ".git",
-    ".git-rewrite",
-    ".hg",
-    ".mypy_cache",
-    ".nox",
-    ".pants.d",
-    ".pytype",
-    ".ruff_cache",
-    ".svn",
-    ".tox",
-    ".venv",
-    "__pypackages__",
-    "_build",
-    "buck-out",
-    "build",
-    "dist",
-    "node_modules",
-    "venv",
-]
 
-#---------------------------------------------------------------------------------------------------
-[tool.ruff.pydocstyle]
-convention = "google"
-
-
-# =============================================== Black ============================================
-[tool.black]
-line-length = 100
+[tool.ruff.lint]
+select = ["ALL"]
+ignore = ["EM101", "EM102", "TRY003", "COM812", "RET504", "S101", "PD008"]
+pydocstyle.convention = "google"
