@@ -9,20 +9,20 @@ import umbridge as ub
 
 # ==================================================================================================
 def process_cli_arguments() -> bool:
-    argParser = argparse.ArgumentParser(
+    arg_parser = argparse.ArgumentParser(
         prog="pto_banana.py",
         usage="python %(prog)s [options]",
         description="Umbridge server-side client emulating PTO map",
     )
 
-    argParser.add_argument(
+    arg_parser.add_argument(
         "-hq",
         "--hyperqueue",
         action="store_true",
         help="Run via Hyperqueue",
     )
 
-    argParser.add_argument(
+    arg_parser.add_argument(
         "-p",
         "--port",
         type=float,
@@ -31,7 +31,7 @@ def process_cli_arguments() -> bool:
         help="User-defined port (if not on Hyperqueue)",
     )
 
-    argParser.add_argument(
+    arg_parser.add_argument(
         "-t",
         "--sleep_times",
         type=float,
@@ -41,10 +41,10 @@ def process_cli_arguments() -> bool:
         help="Sleep times to emulate simulation",
     )
 
-    cliArgs = argParser.parse_args()
-    run_on_hq = cliArgs.hyperqueue
-    local_port = cliArgs.port
-    sleep_times = cliArgs.sleep_times
+    cli_args = arg_parser.parse_args()
+    run_on_hq = cli_args.hyperqueue
+    local_port = cli_args.port
+    sleep_times = cli_args.sleep_times
 
     return run_on_hq, local_port, sleep_times
 
@@ -56,10 +56,10 @@ class PTOModel(ub.Model):
         self._time_coarse, self._time_fine = sleep_times
         self._parameter_ranges = [[500, 2000], [1, 20], [20e9, 30e9], [20e9, 30e9]]
 
-    def get_input_sizes(self, config: dict[str:Any] = {}) -> list[float]:
+    def get_input_sizes(self, _config: dict[str:Any] = {}) -> list[float]:
         return [4]
 
-    def get_output_sizes(self, config: dict[str:Any] = {}) -> list[float]:
+    def get_output_sizes(self, _config: dict[str:Any] = {}) -> list[float]:
         return [4]
 
     def supports_evaluate(self) -> bool:
@@ -92,6 +92,7 @@ class PTOModel(ub.Model):
         observables[1] = np.sqrt(2 * (parameters[1] - 0.25) ** 4)
         observables[2] = np.sqrt(20 * (parameters[3] ** 2 - 2 * parameters[2]) ** 2)
         observables[3] = np.sqrt(2 * (parameters[2] - 0.25) ** 4)
+        observables = [float(observable) for observable in observables]
 
         return [observables]
 
@@ -99,10 +100,7 @@ class PTOModel(ub.Model):
 # ==================================================================================================
 def main():
     run_on_hq, local_port, sleep_times = process_cli_arguments()
-    if run_on_hq:
-        port = int(os.environ["PORT"])
-    else:
-        port = local_port
+    port = int(os.environ["PORT"]) if run_on_hq else local_port
 
     ub.serve_models(
         [
