@@ -8,20 +8,20 @@ import umbridge as ub
 
 # ==================================================================================================
 def process_cli_arguments() -> bool:
-    argParser = argparse.ArgumentParser(
+    arg_parser = argparse.ArgumentParser(
         prog="loglikelihood_gauss.py",
         usage="python %(prog)s [options]",
         description="Umbridge server-side client emulating Gaussian log-likelihood",
     )
 
-    argParser.add_argument(
+    arg_parser.add_argument(
         "-hq",
         "--hyperqueue",
         action="store_true",
         help="Run via Hyperqueue",
     )
 
-    argParser.add_argument(
+    arg_parser.add_argument(
         "-p",
         "--port",
         type=float,
@@ -30,7 +30,7 @@ def process_cli_arguments() -> bool:
         help="User-defined port (if not on Hyperqueue)",
     )
 
-    argParser.add_argument(
+    arg_parser.add_argument(
         "-t",
         "--sleep_times",
         type=float,
@@ -40,10 +40,10 @@ def process_cli_arguments() -> bool:
         help="Sleep times to emulate simulation",
     )
 
-    cliArgs = argParser.parse_args()
-    run_on_hq = cliArgs.hyperqueue
-    local_port = cliArgs.port
-    sleep_times = cliArgs.sleep_times
+    cli_args = arg_parser.parse_args()
+    run_on_hq = cli_args.hyperqueue
+    local_port = cli_args.port
+    sleep_times = cli_args.sleep_times
 
     return run_on_hq, local_port, sleep_times
 
@@ -56,10 +56,10 @@ class GaussianLogLikelihood(ub.Model):
         self._mean = 5e6
         self._covariance = 1e12
 
-    def get_input_sizes(self, config: dict[str, Any] = {}) -> list[int]:
+    def get_input_sizes(self, _config: dict[str, Any] = {}) -> list[int]:
         return [1]
 
-    def get_output_sizes(self, config: dict[str, Any] = {}) -> list[int]:
+    def get_output_sizes(self, _config: dict[str, Any] = {}) -> list[int]:
         return [1]
 
     def supports_evaluate(self) -> bool:
@@ -75,16 +75,13 @@ class GaussianLogLikelihood(ub.Model):
 
         state_diff = parameters[0][0] - self._mean
         log_likelihood = -0.5 * state_diff**2 / self._covariance
-        return [[log_likelihood]]
+        return [[float(log_likelihood)]]
 
 
 # ==================================================================================================
 def main():
     run_on_hq, local_port, sleep_times = process_cli_arguments()
-    if run_on_hq:
-        port = int(os.environ["PORT"])
-    else:
-        port = local_port
+    port = int(os.environ["PORT"]) if run_on_hq else local_port
 
     ub.serve_models(
         [
