@@ -68,8 +68,6 @@ class BaseLogPrior(ABC):
 class UniformLogPrior(BaseLogPrior):
     """Implementation of a uniform prior."""
 
-    
-
     def __init__(self, parameter_intervals: np.ndarray, seed: int) -> None:
         """Constructor.
 
@@ -99,7 +97,7 @@ class UniformLogPrior(BaseLogPrior):
         """
         if not parameter.size == self._interval_lengths.size:
             raise ValueError(f"Invalid parameter dimension {parameter.size}")
-        
+
         has_support = ((parameter >= self._lower_bounds) & (parameter <= self._upper_bounds)).all()
         if has_support:
             return 0
@@ -123,7 +121,7 @@ class GaussianLogPrior(BaseLogPrior):
             mean (np.ndarray): Mean vector
             covariance (np.ndarray): Covariance matrix
             seed (int): RNG seed
-        
+
         Raises:
             ValueError: Checks for valid mean and covariance dimensions
         """
@@ -133,18 +131,19 @@ class GaussianLogPrior(BaseLogPrior):
         self._precision = np.linalg.inv(covariance)
 
         if not mean.size == covariance.shape[0] == covariance.shape[1]:
-            raise ValueError("Mean and covariance dimensions do not match:"
-                             f"{mean.shape}, {covariance.shape}")
+            raise ValueError(
+                f"Mean and covariance dimensions do not match: {mean.shape}, {covariance.shape}"
+            )
 
     def evaluate(self, parameter: np.array) -> float:
         """Compute log-probability for given parameter.
-        
+
         Raises:
             ValueError: Checks for valid parameter dimension
         """
         if not parameter.size == self._mean.size:
             raise ValueError(f"Invalid parameter dimension {parameter.size}")
-        
+
         parameter_diff = parameter - self._mean
         log_probability = -0.5 * parameter_diff.T @ self._precision @ parameter_diff
         return log_probability
