@@ -1,7 +1,7 @@
 """MTMLDA Core Library.
 
 Modules:
-    jobhandling: Threaded submission and retreival of compute jobs, for poserior evaluations
+    jobhandling: Threaded submission and retrieval of compute jobs, for poserior evaluations
     logging: Customized logger for run and debug logs
     mcmc: Markov-Chain Monte Carlo functions, including proposals and accept-reject steps
     mltree: Methods for inspection andmanipulation of Markov Trees
